feat(package.json): initial add of hashticle test project
This is a tentative test project where we create an Saas site to
provide example about how we want to work at lab. To create the best
possible code what we can do now is the most important thing, but the
plan is that creating a viable live site.

We would like to see a good environment where easy to devel, test,
build and deploy.

Signed-off-by: Ádám Gólya <adam.stork@gmail.com>