feat(features/support/tools): initial add of feature tools
This is not the most elegant solution to wait in the test the
async processes, but for now I did not find any better.

In the future it would be awesome if we figure out how can we
remove this, and try to catch that the state is changed and
we have some subscribed function that can be called in this case.